# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Simulação",
        "description": "Documentação da API de Simulação. Esta API fornece funcionalidades para gerenciar clientes, instituições, modalidades de crédito e ofertas.",
        "contact": {
            "name": "Luiz Santos Full Stack Developer ",
            "email": "luizcsdev@gmail.com"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor Local"
        }
    ],
    "paths": {
...
        "/simulacao/credito": {
            "post": {
                "tags": [
                    "Crédito"
                ],
                "summary": "Consulta de crédito",
                "description": "Obtém as instituições financeiras e modalidades de crédito disponíveis para um cliente baseado no CPF.",
                "operationId": "93e4d62df4c13e766f415714abd36d38",
                "requestBody": {
                    "description": "CPF do cliente no formato JSON",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "cpf": {
                                        "type": "string",
                                        "example": "12312312312"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Lista de instituições financeiras e modalidades de crédito disponíveis.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "instituicao_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "instituicao_nome": {
                                                "type": "string",
                                                "example": "Banco Exemplo"
                                            },
                                            "modalidades": {
                                                "type": "array",
                                                "items": {
                                                    "properties": {
                                                        "modalidade_cod": {
                                                            "type": "integer",
                                                            "example": 101
                                                        },
                                                        "modalidade_nome": {
                                                            "type": "string",
                                                            "example": "Crédito Pessoal"
                                                        },
                                                        "taxa_juros": {
                                                            "type": "number",
                                                            "format": "float",
                                                            "example": 1.5
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "CPF não encontrado ou sem ofertas disponíveis.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "CPF não encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/simulacao/simula-oferta": {
            "post": {
                "tags": [
                    "Crédito"
                ],
                "summary": "Simulação de ofertas de crédito",
                "description": "Simula as ofertas de crédito disponíveis para um cliente com base no CPF informado.",
                "operationId": "22aa8d6be9a090ee9f1f9b4bb112eaaf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "cpf"
                                ],
                                "properties": {
                                    "cpf": {
                                        "type": "string",
                                        "example": "11111111111"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ofertas de crédito simuladas.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "cpf": {
                                            "type": "string",
                                            "example": "11111111111"
                                        },
                                        "ofertas": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "instituicaoFinanceira": {
                                                        "type": "string",
                                                        "example": "Banco Exemplo"
                                                    },
                                                    "modalidadeCredito": {
                                                        "type": "string",
                                                        "example": "Crédito Consignado"
                                                    },
                                                    "valorSolicitado": {
                                                        "type": "number",
                                                        "example": 5000
                                                    },
                                                    "valorAPagar": {
                                                        "type": "number",
                                                        "example": 5400
                                                    },
                                                    "taxaJuros": {
                                                        "type": "string",
                                                        "example": "1.2%"
                                                    },
                                                    "qntParcelas": {
                                                        "type": "integer",
                                                        "example": 12
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "CPF não encontrado ou sem ofertas disponíveis.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Nenhuma oferta encontrada para o CPF informado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Erro ao processar a solicitação.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Erro ao processar a solicitação"
                                        },
                                        "error": {
                                            "type": "string",
                                            "example": "Detalhes do erro"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}